backend-python/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
presidio-analyzer
presidio-anonymizer
chromadb
sentence-transformers
openai
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai
